requirements.txt
streamlit==1.38.0
mysql-connector-python
pandas
